restructured and initialized random tensor init
