+++ Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+rand = "0.9.2"
+rand_distr = "0.5.1"
