+++ src/lib.rs
@@ -1,6 +1,9 @@
 pub mod tensor {
     pub mod core;
-    pub mod operations;
+    pub mod ops;
+    pub mod reductions;
+    pub mod nn;
+    pub mod init;
 }
 pub mod utils{
     pub mod float;
